requirements.txt
PyYAML
sense-hat
boto3
datetime
os
pandas
pathlib
requests
python-dotenv